refactor(storage): isolate KV root decoding
