+++ docs/astrid-principal-store-operations.md
@@ -128,18 +128,14 @@ ChunkingProfile {
 ```
 
 Content-defined chunking avoids shifting every later boundary after an
-insertion near the start of a file. The first implemented profile pins the
-FastCDC 2020 algorithm and implementation revision, normalization level one, an
-unseeded gear table, 16 KiB minimum, 64 KiB target, and 256 KiB maximum. Every
-field is encoded into the file object, the dependency is exact-pinned, and
-golden cut points prevent an ordinary dependency update from changing stored
-identity.
-
-The profile is a compatibility point, not a claim of universal optimality.
-Astrid must still benchmark source trees, package caches, model artifacts,
-databases, VM images, compressed media, encrypted data, and adversarial random
-data before selecting additional profiles. A different profile reconstructs
-the same bytes but deliberately produces a different file identity.
+insertion near the start of a file. The first profile pins FastCDC 2020,
+implementation revision, normalization level one, an unseeded gear table, and
+16/64/256 KiB minimum/target/maximum sizes. Every field is encoded into the
+file object; an exact dependency pin and golden cuts protect stored identity.
+This compatibility point is not universally optimal. Astrid must benchmark
+source trees, caches, model artifacts, databases, VM images, compressed or
+encrypted media, and adversarial random data before adding profiles. Another
+profile reconstructs the same bytes with a deliberately different file ID.
 
 ### 14.3 Collision probability
